add passed tests

+++ homework4try2/src/main/java/com/a1arick/spbsu/homework4/try2/server/model/GameModel.java
@@ -12,7 +12,7 @@ public class GameModel extends AbstractGameModel{
     }
 
     @Override
-    protected long getTime() {
+    protected double getTime() {
         return System.currentTimeMillis();
     }
 }
